# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Allam Journal API",
    "version": "1.0.0",
    "description": "Single-journal OJS workflow. JSON commands and multipart uploads require a session-bound CSRF token. See README for login and resource policies. Response bodies are projections documented in architecture.md."
  },
  "servers": [
    {
      "url": "/",
      "description": "The backend serving this documentation"
    }
  ],
  "paths": {
...
    "/api/covers": {
      "post": {
        "summary": "POST /api/covers",
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Permission or CSRF denied"
          },
          "409": {
            "description": "Workflow, version, or uniqueness conflict"
          },
          "201": {
            "description": "Created resource with id"
          }
        },
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-TOKEN",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Fetch /api/auth/csrf using the same session cookie."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "tags": [
          "Files & galleys"
        ],
        "operationId": "post_api_covers"
      }
    },
    "/api/issues/{id}/cover": {
      "put": {
        "summary": "PUT /api/issues/{id}/cover",
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Permission or CSRF denied"
          },
          "409": {
            "description": "Workflow, version, or uniqueness conflict"
          },
          "204": {
            "description": "Successful command"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-TOKEN",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Fetch /api/auth/csrf using the same session cookie."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoverInput"
              }
            }
          }
        },
        "tags": [
          "Issues & publication"
        ],
        "operationId": "put_api_issues_id_cover"
      }
    },
    "/api/public/covers/{id}": {
      "get": {
        "summary": "GET /api/public/covers/{id}",
        "security": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Permission or CSRF denied"
          },
          "409": {
            "description": "Workflow, version, or uniqueness conflict"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Public journal"
        ],
        "operationId": "get_api_public_covers_id"
      }
    },
...
    "/api/preservation/manifest": {
      "get": {
        "summary": "GET /api/preservation/manifest",
        "security": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Permission or CSRF denied"
          },
          "409": {
            "description": "Workflow, version, or uniqueness conflict"
          }
        },
        "tags": [
          "Preservation"
        ],
        "operationId": "get_api_preservation_manifest"
      }
    },
    "/api/preservation/issues/{id}/manifest": {
      "get": {
        "summary": "GET /api/preservation/issues/{id}/manifest",
        "security": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Permission or CSRF denied"
          },
          "409": {
            "description": "Workflow, version, or uniqueness conflict"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Preservation"
        ],
        "operationId": "get_api_preservation_issues_id_manifest"
      }
    },
    "/api/preservation/issues/{id}/inventory": {
      "get": {
        "summary": "GET /api/preservation/issues/{id}/inventory",
        "security": [],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Permission or CSRF denied"
          },
          "409": {
            "description": "Workflow, version, or uniqueness conflict"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Preservation"
        ],
        "operationId": "get_api_preservation_issues_id_inventory"
      }
    },
    "/api/preservation/issues/{id}/package.zip": {
      "get": {
        "summary": "GET /api/preservation/issues/{id}/package.zip",
        "security": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/zip": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Permission or CSRF denied"
          },
          "409": {
            "description": "Workflow, version, or uniqueness conflict"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Preservation"
        ],
        "operationId": "get_api_preservation_issues_id_package_zip"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "JSESSIONID"
      }
    },
    "schemas": {
...
      "CoverInput": {
        "type": "object",
        "properties": {
          "coverId": {
            "type": "string"
          }
        },
        "required": [
          "coverId"
        ]
      }
    }
  },
...
}